# Hestia-Swagger/swagger.yml
openapi: 3.0.3
info:
  title: Hestia — API Reference
  description: |
    Aggregated API reference for all Hestia services.
    Each path overrides the server URL to route to the correct service.

    | Service   | Port  |
    |-----------|-------|
    | Hub       | 19001 |
    | Archive   | 19002 |
    | Hecate    | 19003 |
    | Oracle    | 19004 |
    | Hermes    | 19005 |
    | Scout     | 19006 |
    | Chronos   | 19007 |
    | Argus     | 19008 |
    | Iris      | 19012 |
    | Hephaestus| 19010 |
    | Dummy     | 19011 |
  version: 1.0.0
servers:
  - url: http://localhost:19001
    description: Hub (default)
...
components:
  schemas:
...
    ModuleToolQueryRequest:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
          example: real_estate
        query:
          type: string
        session_id:
          type: string
        limit:
          type: integer
          default: 20
          minimum: 1
          maximum: 50
        filters:
          type: object
          additionalProperties: true
        filters_gt:
          type: object
          additionalProperties: true
        filters_lt:
          type: object
          additionalProperties: true
        sort_by:
          type: string
        sort_order:
          type: string
          enum: [asc, desc]
        preferences:
          type: array
          items:
            type: object
    RealEstateSearchRequest:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 20
          minimum: 1
          maximum: 50
        city:
          type: string
        nearby:
          type: string
        radius_km:
          type: number
        price_max:
          type: number
        price_min:
          type: number
        rooms_min:
          type: integer
        surface_min:
          type: number

# ─────────────────────────────────────────────────────────────────────────────
#  Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/logs:
    get:
      tags: [Hub]
      summary: Retrieve buffered runtime logs (service-local endpoint)
      description: |
        Implemented in all runtime services.
        Current instances:
        - Hub (19001)
        - Archive (19002)
        - Hecate (19003)
        - Oracle (19004)
        - Hermes (19005)
        - Scout (19006)
        - Chronos (19007)
        - Argus (19008)
        - Athena (19009)
        - Hephaestus (19010)
        - Telegram control API (internal: hestia_telegram:8010)

        Sensitive values (tokens, passwords, secrets, credential URLs, auth headers)
        are redacted before entries are buffered and returned.

        Use query filters to keep output dense and useful for troubleshooting.
      servers:
        - url: http://localhost:19001
          description: Hub
        - url: http://localhost:19002
          description: Archive
        - url: http://localhost:19003
          description: Hecate
        - url: http://localhost:19012
          description: Iris
        - url: http://localhost:19004
          description: Oracle
        - url: http://localhost:19005
          description: Hermes
        - url: http://localhost:19006
          description: Scout
        - url: http://localhost:19007
          description: Chronos
        - url: http://localhost:19008
          description: Argus
        - url: http://localhost:19009
          description: Athena
        - url: http://localhost:19010
          description: Hephaestus
        - url: http://hestia_telegram:8010
          description: Telegram control API (Docker network)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            maximum: 2000
        - name: level
          in: query
          schema:
            type: string
            example: INFO
        - name: contains
          in: query
          schema:
            type: string
            example: outbound_event_id=p2-4-smoke-3
      responses:
        "200":
          description: "{service, count, logs[]}"
...
  /api/module-tools/query:
    post:
      tags: [Scout]
      summary: Query entities via module-tool interface
      servers:
        - url: http://localhost:19006
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleToolQueryRequest"
      responses:
        "200":
          description: "{domain, items: [...]}"
...
  /api/tools/real_estate/search:
    post:
      tags: [Scout]
      summary: Search real estate listings
      servers:
        - url: http://localhost:19006
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RealEstateSearchRequest"
      responses:
        "200":
          description: List of real estate entity items

  # ══════════════════════════════════════════════════════════════════════════
  #  CHRONOS — 19007
  # ══════════════════════════════════════════════════════════════════════════
...
  /api/module/maintenance/reconcile:
    post:
      tags: [Maintenance]
      summary: Execute module maintenance reconcile using the standard contract
      servers:
        - url: http://localhost:19002
        - url: http://localhost:19006
        - url: http://localhost:19007
        - url: http://localhost:19011
        - url: http://localhost:19012
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                task_id:
                  type: string
                  nullable: true
                issue:
                  type: string
                  nullable: true
                requested_action:
                  type: string
                  nullable: true
                environment:
                  type: string
                  default: dev
                dry_run:
                  type: boolean
                  default: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: "{status, service, dry_run, task_id, executed_at, retriable, summary, mutation_count, details}"
  /api/maintenance/reconcile:
    post:
      tags: [Maintenance]
      summary: Compatibility alias of module maintenance reconcile
      servers:
        - url: http://localhost:19002
        - url: http://localhost:19006
        - url: http://localhost:19007
        - url: http://localhost:19011
        - url: http://localhost:19012
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: "Alias response; same as /api/module/maintenance/reconcile"

  # ══════════════════════════════════════════════════════════════════════════
  #  ATHENA — 19009
  # ══════════════════════════════════════════════════════════════════════════
...